feat: introduce AWS_VAULT_STDOUT as env var to configure the --stdout flag
Sometimes it's tedious to add `--stdout` on all invocations of
`aws-vault`, this allows it to be configured as an evironment variable
for all times.

Especially useful when your main browser is not the one that's logged in
to AWS (I'm shuffling dozens of aws accounts AND browsers at the same
time).